remove usb_bus_sense warning

+++ firmware/u4a2/system/usb/usb9/usb9.c
@@ -226,10 +226,10 @@ void USBStdGetStatusHandler(void)
              * _byte0: bit0: Self-Powered Status [0] Bus-Powered [1] Self-Powered
              *         bit1: RemoteWakeup        [0] Disabled    [1] Enabled
              */
-            if(self_power == 1)                     // self_power defined in io_cfg.h
+            if(self_power == (unsigned) 1)          // self_power defined in io_cfg.h
                 CtrlTrfData._byte0|=0b000000001;    // Set bit0
 
-            if(usb_stat.RemoteWakeup == (unsigned) 1)          // usb_stat defined in usbmmap.c
+            if(usb_stat.RemoteWakeup == (unsigned) 1) // usb_stat defined in usbmmap.c
                 CtrlTrfData._byte0|=0b00000010;     // Set bit1
             break;
         case RCPT_INTF:
